fix: improve compiler compatibility and warnings (#647)
* fix: improve compatibility checks for atomic operations

- Replace GNUC macro check with __GNUC_PREREQ(4, 1) for more precise version control
- Add fallback definition for __GNUC_PREREQ macro when not available

* fix: add explicit type casting in gettick_ms()

- Add explicit (unsigned int) casting to tv_sec and tv_nsec/tv_usec calculations to prevent potential integer overflow

* fix: suppress compiler warnings for unused variable and function

- Add (void) cast and inline qualifier to eliminate unused warnings

---------

+++ base/htime.c
@@ -15,11 +15,11 @@ unsigned int gettick_ms() {
 #elif HAVE_CLOCK_GETTIME
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+    return (unsigned int)ts.tv_sec * 1000 + (unsigned int)ts.tv_nsec / 1000000;
 #else
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    return (unsigned int)tv.tv_sec * 1000 + (unsigned int)tv.tv_usec / 1000;
 #endif
 }
